Fix the java sample and update the Makefile.
We need to make the gyp build generate one static library file for the
fletch VM. For now I just hacked that into the Makefile, but will look
into getting that into the gyp build instead.

R=kasperl@google.com

Review URL: https://codereview.chromium.org/904843003

+++ tests/service_tests/echo/java/Echo.java
@@ -56,13 +56,13 @@ class Echo {
       }
     });
     System.out.println("Java: async echo call with argument " + 3);
-    EchoService.SumAsync(3, 4, new EchoService.SumCallback() {
+    EchoService.SumAsync((short) 3, 4, new EchoService.SumCallback() {
       public void handle(int result) {
         System.out.println("Java: async sum result " + result);
       }
     });
     System.out.println("Java: async sum call with arguments " + 3 + " " + 4);
-    result = EchoService.Sum(3, 4);
+    result = EchoService.Sum((short) 3, 4);
     System.out.println("Java: result of sum(3, 4) is " + result);
 
     EchoService.TearDown();
